fix an overflow error in igraph_static_fitness_game(), closes #1094

+++ src/games.c
@@ -3578,14 +3578,14 @@ int igraph_static_fitness_game(igraph_t *graph, igraph_integer_t no_of_edges,
         if (VECTOR(*fitness_out)[i] != 0 && VECTOR(*fitness_in)[i] != 0)
             nodes++;
       }
-      max_no_of_edges = outnodes * innodes - (loops ? 0 : nodes);
+      max_no_of_edges = ((long int) outnodes) * innodes - (loops ? 0 : nodes);
     } else {
       nodes = 0;
       for (i=0; i < no_of_nodes; i++) {
         if (VECTOR(*fitness_out)[i] != 0)
           nodes++;
       }
-      max_no_of_edges = loops ? nodes*(nodes+1)/2 : nodes*(nodes-1)/2;
+      max_no_of_edges = loops ? nodes*((long int)nodes+1)/2 : nodes*((long int)nodes-1)/2;
     }
     if (no_of_edges > max_no_of_edges)
       IGRAPH_ERROR("Too many edges requested", IGRAPH_EINVAL);
